Java 25; Gradle 9.4.0; dependency version bumps; fix deprecation warnings and test error

+++ CSVParser/src/main/java/org/avalancs/csvparser/FieldTypes.java
@@ -11,7 +11,7 @@ public class FieldTypes {
     public static final FieldType NUMBER = new FieldType("Number") {
         @Override
         public Object parse(String col) {
-            return Integer.parseInt(col);
+            return Double.parseDouble(col);
         }
     };
